requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
certifi==2024.2.2
cssselect==1.2.0
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
jmespath==1.0.1
lxml==5.2.1
magic-filter==1.0.12
multidict==6.0.5
packaging==24.0
parsel==1.9.1
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
python-decouple==3.8
python-dotenv==1.0.1
sniffio==1.3.1
tkcalendar==1.6.1
ttkbootstrap==1.10.1
ttkwidgets==0.13.0
typing_extensions==4.10.0
w3lib==2.1.2
yarl==1.9.4

requests~=2.31.0
bs4~=0.0.2
beautifulsoup4~=4.12.3